Count time in samples
git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@1556 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GOSoundSampler.h
@@ -36,7 +36,7 @@ typedef struct GO_SAMPLER_T
 	unsigned                   audio_group_id;
 	audio_section_stream       stream;
 	GOSoundFader               fader;
-	unsigned                   time;
+	uint64_t                   time;
 	unsigned                   velocity;
 	unsigned                   delay;
 	/* current index of the current block into this sample */
